Issue #9: Add Pods/.
The trade-offs between adding Pods/ to VC & ignoring it are listed in:
  http://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control

I would prefer not to add it because I'm concerned about:
* conflicts from merging in VC
* I'm not sure but I suspect that in the current config, projects that use this
  lib download our dependencies *and* their own dependencies, even if they're
  the same - could they be shared if we didn't include the deps in our repo?

That being said, it appears this is the desired configuration while using
CocoaPods with Carthage: the _Pods.xcodeproj link exists in the template for
use with Carthage. It points at the Pods/ project (expecting to find the
ScrollingCardView scheme there), meaning that the Pods/ project needs to exist
when Carthage runs. If we don't have Pods/ checked in, Carthage would have to
generate them by using CocoaPods, which defeats the purpose of using Carthage.
Thus, we need to pre-generate Pods/ and check it into VC.

Since our Pods/ dependencies all exist for testing, not production, we might be
able to link this to our regular .xcodeproj, but I didn't want to spend the
time to figure it out.

---
fwiw, this project was generated from `pod lib create` - I wonder if we could
come up with a more logical and simple project design if we started from
scratch. Perhaps https://goo.gl/mzZ3Si can be of help?